+++ src/PCA9685.cpp
@@ -43,7 +43,7 @@ void PCA9685::setChannelPulse(size_t leg, size_t channel, float pulse_ms) {
     setServoPulse(pulsed, pulse_ms);
 }
 
-void PCA9685::addLeg(uint8_t ch0, uint8_t ch1, uint8_t ch2) {
+void PCA9685::addLeg(size_t ch0, size_t ch1, size_t ch2) {
     legs.emplace_back(RobotLeg(ch0, ch1, ch2));
 }
